+++ patzilla-ui/navigator/app/config.js
@@ -51,6 +51,9 @@ NavigatorConfiguration = Backbone.Model.extend({
             baseurl += path;
         }
 
+        // Strip trailing slashes
+        baseurl = _.string.rtrim(baseurl, '/');
+
         this.set('baseurl', baseurl);
 
     },