+++ patzilla/navigator/__init__.py
@@ -1,62 +1,11 @@
 # -*- coding: utf-8 -*-
 # (c) 2013-2017 Andreas Motl, Elmyra UG <andreas.motl@elmyra.de>
-from ConfigParser import NoOptionError
-from patzilla.util.config import read_config
-from patzilla.util.data.container import SmartBunch
-from patzilla.version import __VERSION__
-from patzilla.navigator.util import PngRenderer, XmlRenderer, PdfRenderer, NullRenderer
-from pyramid.config import Configurator
-
-def main(global_config, **settings):
-    """This function returns a Pyramid WSGI application."""
-
-    settings.setdefault('CONFIG_FILE', global_config.get('__file__'))
-    settings.setdefault('SOFTWARE_VERSION', __VERSION__)
-
-    application_settings = read_config(settings['CONFIG_FILE'], kind=SmartBunch)
-    try:
-        datasources = application_settings.ip_navigator.datasources
-    except:
-        raise NoOptionError('datasources', 'ip_navigator')
-
-    config = Configurator(settings=settings)
-    config.registry.application_settings = application_settings
-
-    # Addons
-    config.include('pyramid_mako')
-    config.include('pyramid_beaker')
-
-    config.scan('patzilla.util.web.pyramid.cornice')
-    config.include('cornice')
-    #config.include("akhet.static")
-
-    # Register subsystem components
-    config.include("patzilla.util.web.identity")
-    config.include("patzilla.util.database.beaker_mongodb_gridfs")
-    config.include("patzilla.util.web.pyramid")
+def includeme(config):
 
     # Register application components: URL generator, renderer globals
     config.include("patzilla.navigator.opaquelinks")
     config.include("patzilla.navigator.subscribers")
-    config.include("patzilla.navigator.views")
-
-    if 'ops' in datasources:
-        config.include("patzilla.access.epo.ops.client")
-
-    if 'depatisconnect' in datasources:
-        config.include("patzilla.access.dpma.depatisconnect")
-
-    if 'ificlaims' in datasources:
-        config.include("patzilla.access.ificlaims.clientpool")
-
-    if 'depatech' in datasources:
-        config.include("patzilla.access.depatech.clientpool")
-
-    config.add_mako_renderer('.html')
-    config.add_renderer('xml', XmlRenderer)
-    config.add_renderer('png', PngRenderer)
-    config.add_renderer('pdf', PdfRenderer)
-    config.add_renderer('null', NullRenderer)
+    config.include("patzilla.navigator.views", route_prefix='/navigator')
 
     # Views and routes
     config.add_static_view('static', 'static', cache_max_age=3600)
@@ -64,5 +13,3 @@ def main(global_config, **settings):
     config.add_route('home', '/')
 
     config.scan()
-
-    return config.make_wsgi_app()